Organize Selector and its children
